Added the event-core service to allow pulling events
