+++ .angular/cache/17.1.3/vite/deps/@angular_common.js
@@ -96,10 +96,11 @@ import {
   provideImgixLoader,
   registerLocaleData,
   setRootDomAdapter
-} from "./chunk-OY2YHKMJ.js";
+} from "./chunk-IGU6QSFS.js";
 import {
   IMAGE_CONFIG
-} from "./chunk-JEG2PW3Z.js";
+} from "./chunk-Z34T3VMF.js";
+import "./chunk-SXIXOCJ4.js";
 export {
   APP_BASE_HREF,
   AsyncPipe,
