+++ .angular/cache/17.1.3/vite/deps/@angular_core.js
@@ -432,7 +432,8 @@ import {
   ɵɵtrustConstantResourceUrl,
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery
-} from "./chunk-JEG2PW3Z.js";
+} from "./chunk-Z34T3VMF.js";
+import "./chunk-SXIXOCJ4.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
