+++ .angular/cache/17.1.3/vite/deps/@angular_platform-browser.js
@@ -35,11 +35,12 @@ import {
   provideProtractorTestingSupport,
   withHttpTransferCacheOptions,
   withNoHttpTransferCache
-} from "./chunk-GIFUNVID.js";
+} from "./chunk-7IK53PET.js";
 import {
   getDOM
-} from "./chunk-OY2YHKMJ.js";
-import "./chunk-JEG2PW3Z.js";
+} from "./chunk-IGU6QSFS.js";
+import "./chunk-Z34T3VMF.js";
+import "./chunk-SXIXOCJ4.js";
 export {
   BrowserModule,
   By,
