+++ .angular/cache/17.1.3/vite/deps/_metadata.json
@@ -1,49 +1,58 @@
 {
-  "hash": "123a019a",
-  "configHash": "fa0992fe",
+  "hash": "78d50ce2",
+  "configHash": "ee1c459d",
   "lockfileHash": "a61bbb0b",
-  "browserHash": "e309975b",
+  "browserHash": "b2173105",
   "optimized": {
     "@angular/common": {
       "src": "../../../../../node_modules/@angular/common/fesm2022/common.mjs",
       "file": "@angular_common.js",
-      "fileHash": "a59b9a38",
+      "fileHash": "1e261cbb",
       "needsInterop": false
     },
     "@angular/core": {
       "src": "../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "c2f4d419",
+      "fileHash": "27749ae6",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "baeedac2",
+      "fileHash": "145c5420",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "73915cd6",
+      "fileHash": "a8dd13d2",
       "needsInterop": false
     },
     "@ng-bootstrap/ng-bootstrap": {
       "src": "../../../../../node_modules/@ng-bootstrap/ng-bootstrap/fesm2022/ng-bootstrap.mjs",
       "file": "@ng-bootstrap_ng-bootstrap.js",
-      "fileHash": "e44882f6",
+      "fileHash": "021951d9",
+      "needsInterop": false
+    },
+    "rxjs": {
+      "src": "../../../../../node_modules/rxjs/dist/esm5/index.js",
+      "file": "rxjs.js",
+      "fileHash": "1b8b0935",
       "needsInterop": false
     }
   },
   "chunks": {
-    "chunk-GIFUNVID": {
-      "file": "chunk-GIFUNVID.js"
+    "chunk-7IK53PET": {
+      "file": "chunk-7IK53PET.js"
     },
-    "chunk-OY2YHKMJ": {
-      "file": "chunk-OY2YHKMJ.js"
+    "chunk-IGU6QSFS": {
+      "file": "chunk-IGU6QSFS.js"
     },
-    "chunk-JEG2PW3Z": {
-      "file": "chunk-JEG2PW3Z.js"
+    "chunk-Z34T3VMF": {
+      "file": "chunk-Z34T3VMF.js"
+    },
+    "chunk-SXIXOCJ4": {
+      "file": "chunk-SXIXOCJ4.js"
     }
   }
 }
